Some drag drop styling fixes
Still not ideal, as the drag doesn't end until the mouse moves after
mouse release, as we seemingly can't detect global mouse release events.

So if the mouse is not on the window list when released after a drag,
the button incorrectly retains its hover state until the mouse moves.