Move styling into separate css file, and refactor drag-and-drop reordering
drag and drop now uses a timer that polls the mouse state for button
release and position, so that we can detect mouse release to end the
drag without relying on movement. Further, we call the same callback
when the mouse enters another button, to prevent flicker changing
styles before the timer fires only to change them again.