Actually fix dash to panel enable/disable/reset detection
It turns out that whenever the user disables an extension that was
enabled prior to  dash to panel, GNOME shell resets dash to panel - it
calls dash to panel's `disable()` and `enable()` methods, which results
in a new `global.dashToPanel` object. However, GNOME shell does *not*
emit events to inform listeners of this as an extension changing state.

To deal with this, we just constantly check whenever any extension
changes state, whether the `global.dashToPanel` object has appeared,
disappeared, or been replaced, and respond accordingly.