Fix drag cancellation when dragging buttons for maximised windows between monitors
Maximised windows don't move right away due to the need to avoid some
race conditions, so we need to allow for this to be delayed. This
results in the window-entered-monitor signal being emitted, which we
use to update our button visibility. Previously any change in dragged
button visibility would cancel the drag, now we only cancel if the
change is from visible to hidden, not from hidden to visible.